Move unit test and add navigation
