+++ lib/tests/questionlib_test.php
@@ -23,7 +23,6 @@
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 use core_tag\output\tag;
-use qbank_managecategories\managecategories_helper;
 
 
 defined('MOODLE_INTERNAL') || die();
@@ -516,68 +515,6 @@ class core_questionlib_testcase extends advanced_testcase {
 
         // Verify there is just a single top-level category.
         $this->assertEquals(1, $DB->count_records('question_categories', ['contextid' => $qcat->contextid, 'parent' => 0]));
-    }
-
-    public function test_question_remove_stale_questions_from_category() {
-        global $DB;
-        $this->resetAfterTest(true);
-        $this->setAdminUser();
-
-        $dg = $this->getDataGenerator();
-        $course = $dg->create_course();
-        $quiz = $dg->create_module('quiz', ['course' => $course->id]);
-
-        $qgen = $dg->get_plugin_generator('core_question');
-        $context = context_system::instance();
-
-        $qcat1 = $qgen->create_question_category(['contextid' => $context->id]);
-        $q1a = $qgen->create_question('shortanswer', null, ['category' => $qcat1->id]);     // Will be hidden.
-        $DB->set_field('question', 'hidden', 1, ['id' => $q1a->id]);
-
-        $qcat2 = $qgen->create_question_category(['contextid' => $context->id]);
-        $q2a = $qgen->create_question('shortanswer', null, ['category' => $qcat2->id]);     // Will be hidden.
-        $q2b = $qgen->create_question('shortanswer', null, ['category' => $qcat2->id]);     // Will be hidden but used.
-        $DB->set_field('question', 'hidden', 1, ['id' => $q2a->id]);
-        $DB->set_field('question', 'hidden', 1, ['id' => $q2b->id]);
-        quiz_add_quiz_question($q2b->id, $quiz);
-        quiz_add_random_questions($quiz, 0, $qcat2->id, 1, false);
-
-        // We added one random question to the quiz and we expect the quiz to have only one random question.
-        $q2d = $DB->get_record_sql("SELECT q.*
-                                      FROM {question} q
-                                      JOIN {quiz_slots} s ON s.questionid = q.id
-                                     WHERE q.qtype = :qtype
-                                           AND s.quizid = :quizid",
-                array('qtype' => 'random', 'quizid' => $quiz->id), MUST_EXIST);
-
-        // The following 2 lines have to be after the quiz_add_random_questions() call above.
-        // Otherwise, quiz_add_random_questions() will to be "smart" and use them instead of creating a new "random" question.
-        $q1b = $qgen->create_question('random', null, ['category' => $qcat1->id]);          // Will not be used.
-        $q2c = $qgen->create_question('random', null, ['category' => $qcat2->id]);          // Will not be used.
-
-        $this->assertEquals(2, $DB->count_records('question', ['category' => $qcat1->id]));
-        $this->assertEquals(4, $DB->count_records('question', ['category' => $qcat2->id]));
-
-        // Non-existing category, nothing will happen.
-        managecategories_helper::question_remove_stale_questions_from_category(0);
-        $this->assertEquals(2, $DB->count_records('question', ['category' => $qcat1->id]));
-        $this->assertEquals(4, $DB->count_records('question', ['category' => $qcat2->id]));
-
-        // First category, should be empty afterwards.
-        managecategories_helper::question_remove_stale_questions_from_category($qcat1->id);
-        $this->assertEquals(0, $DB->count_records('question', ['category' => $qcat1->id]));
-        $this->assertEquals(4, $DB->count_records('question', ['category' => $qcat2->id]));
-        $this->assertFalse($DB->record_exists('question', ['id' => $q1a->id]));
-        $this->assertFalse($DB->record_exists('question', ['id' => $q1b->id]));
-
-        // Second category, used questions should be left untouched.
-        managecategories_helper::question_remove_stale_questions_from_category($qcat2->id);
-        $this->assertEquals(0, $DB->count_records('question', ['category' => $qcat1->id]));
-        $this->assertEquals(2, $DB->count_records('question', ['category' => $qcat2->id]));
-        $this->assertFalse($DB->record_exists('question', ['id' => $q2a->id]));
-        $this->assertTrue($DB->record_exists('question', ['id' => $q2b->id]));
-        $this->assertFalse($DB->record_exists('question', ['id' => $q2c->id]));
-        $this->assertTrue($DB->record_exists('question', ['id' => $q2d->id]));
     }
 
     /**
