+++ question/bank/managecategories/classes/navigation.php
@@ -1,0 +1,60 @@
+<?php
+// This file is part of Moodle - http://moodle.org/
+//
+// Moodle is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Moodle is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
+
+/**
+ * Plugin entrypoint for navigation.
+ *
+ * @package    qbank_managecategories
+ * @copyright  2021 Catalyst IT Australia Pty Ltd
+ * @author     Guillermo Gomez Arias <guillermogomez@catalyst-au.net>
+ * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+
+namespace qbank_managecategories;
+
+defined('MOODLE_INTERNAL') || die();
+
+/**
+ * Class navigation.
+ *
+ * @package    qbank_managecategories
+ * @copyright  2021 Catalyst IT Australia Pty Ltd
+ * @author     Guillermo Gomez Arias <guillermogomez@catalyst-au.net>
+ * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+class navigation extends \core_question\local\bank\navigation_plugin_base {
+
+    /**
+     * Get export plugin title for this node.
+     */
+    public function get_title(): string {
+        return get_string('categories', 'question');
+    }
+
+    /**
+     * Get export plugin key for this node.
+     */
+    public function get_navigation_key(): string {
+        return 'categories';
+    }
+
+    /**
+     * Get export plugin URL for this node.
+     */
+    public function get_url(): \moodle_url {
+        return new \moodle_url('/question/bank/managecategories/category.php');
+    }
+}
